+++ config/routes/data.rb
@@ -173,9 +173,6 @@ end
 
 resources :contents do
   concerns [:data_routes]
-  collection do
-    get :filter
-  end
 end
 
 resources :controlled_vocabulary_terms do