+++ lib/queries/content/filter.rb
@@ -11,21 +11,34 @@ module Queries
       # @param otu_id [Array, Integer, String, nil]
       attr_accessor :otu_id
 
+      # @return [Array]
+      # @param project_id [Array, Integer, String, nil]
+      attr_accessor :project_id
+
+      # @return [Boolean, nil]
+      attr_accessor :exact
+
       attr_accessor :text # was query_string
 
-      # attr_accessor :most_recent_updates
+      # @return [Boolean, nil]
+      #   true - only content with depictions
+      attr_accessor :depictions
+
+      # @return [Boolean, nil]
+      #   true - only content with citations
+      attr_accessor :citations
 
       # @param [Hash] args
       def initialize(params)
 
-        # topic_id: nil, otu_id: nil, hours_ago: nil, query_string: nil, most_recent_updates: nil
-
         @topic_id = params[:topic_id]
         @otu_id = params[:otu_id]
+        @project_id = params[:project_id]
         @text = params[:text]
-
-       # @hours_ago = hours_ago.to_i if hours_ago
-       # @most_recent_updates = most_recent_updates.to_i
+        @exact = (params[:exact]&.downcase == 'true' ? true : false) if !params[:exact].nil?
+        @depictions = (params[:depictions]&.downcase == 'true' ? true : false) if !params[:depictions].nil?
+      
+        @citations = (params[:citations]&.downcase == 'true' ? true : false) if !params[:citations].nil?
       end
 
       def topic_id
@@ -34,6 +47,10 @@ module Queries
 
       def otu_id
         [@otu_id].flatten.compact
+      end
+
+      def project_id
+        [@project_id].flatten.compact
       end
 
       # @return [Arel::Table]
@@ -46,7 +63,8 @@ module Queries
         clauses = [
           text_facet,
           topic_id_facet,
-          otu_id_facet
+          otu_id_facet,
+          project_id_facet,
         ].compact
 
         a = clauses.shift
@@ -58,7 +76,12 @@ module Queries
 
       # @return [Arel::Node, nil]
       def text_facet
-        text.blank? ? nil : table[:text].eq(text) 
+        return nil if text.blank?
+        if exact
+          table[:text].eq(text.strip)
+        else
+          table[:text].matches('%' + text.strip.gsub(/\s+/, '%') + '%')
+        end
       end
 
       # @return [Arel::Node, nil]
@@ -68,24 +91,73 @@ module Queries
       end
 
       # @return [Arel::Node, nil]
+      def project_id_facet
+        return nil if project_id.empty?
+        table[:project_id].eq_any(project_id) 
+      end
+
+      # TODO: DRY depictions/citations
+
+      # @return [Arel::Node, nil]
+      def depictions_facet
+        return nil if depictions.nil?
+        if depictions
+          ::Content.joins(:depictions)
+        else
+          ::Content.left_joins(:depictions).where(depictions: {id: nil})
+        end
+      end
+
+      # @return [Arel::Node, nil]
+      def citations_facet
+        return nil if citations.nil?
+        if citations
+          ::Content.joins(:citations)
+        else
+          ::Content.left_joins(:citations).where(citations: {id: nil})
+        end
+      end
+
+      # @return [Arel::Node, nil]
       def topic_id_facet
         return nil if topic_id.empty?
         table[:topic_id].eq_any(topic_id) 
       end
 
-      # @return [ActiveRecord::Relation]
-      def all
-        if a = and_clauses
-          ::Content.includes(:otu, :topic).where(and_clauses)
-        else
-          ::Content.includes(:otu, :topic).all
+      def merge_clauses
+        clauses = [
+          depictions_facet, 
+          citations_facet,
+        ].compact
+
+        return nil if clauses.empty?
+
+        a = clauses.shift
+        clauses.each do |b|
+          a = a.merge(b)
         end
+        a
       end
 
-      # @return [Arel::Nodes::Equatity]
-      def recent
-        table[:updated_at].gt(hours_ago.hours.ago) if hours_ago
+      # @return [ActiveRecord::Relation]
+      def all
+        a = and_clauses
+        b = merge_clauses
+
+        q = nil
+        if a && b
+          q = b.where(a)
+        elsif a
+          q = ::Content.includes(:otu, :topic).where(a)
+        elsif b
+          q = b
+        else
+          q = ::Content.includes(:otu, :topic).all
+        end
+
+        q
       end
+
     end
 
   end